BeigeBox/StegaStamp/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers==0.30.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.60.0
scikit-image>=0.19.0
tensorflow>=2.10.0
lpips>=0.1.4

